# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UniTracker API",
        "description": "L'API UniTracker permet de suivre en temps réel les bus universitaires, planifier les trajets, et recevoir des notifications intelligentes. Elle facilite la gestion des transports pour les étudiants et les chauffeurs via des requêtes HTTP sécurisées.",
        "termsOfService": "https://unitracker.ucbc.cd/terms",
        "contact": {
            "name": "Équipe Support UniTracker",
            "email": "support@unitracker.cd"
        },
        "license": {
            "name": "UCBC / UniTracker",
            "url": "https://unitracker.ucbc.cd/license"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://unitracker.ucbc.cd",
            "description": "Serveur de production"
        }
    ],
    "paths": {
        "/api/buses": {
            "get": {
                "tags": [
                    "Bus"
                ],
                "summary": "Lister tous les bus",
                "operationId": "f519550ca830b5dcb346e18781ed50b9",
                "responses": {
                    "200": {
                        "description": "Liste des bus avec chauffeurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "buses": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "immatriculation": {
                                                        "type": "string"
                                                    },
                                                    "marque": {
                                                        "type": "string"
                                                    },
                                                    "modele": {
                                                        "type": "string"
                                                    },
                                                    "couleur": {
                                                        "type": "string"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "chauffeur_id": {
                                                        "type": "integer"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bus"
                ],
                "summary": "Créer un nouveau bus",
                "operationId": "e1409a5d1e40f031070c6614524fe308",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "immatriculation",
                                    "chauffeur_id"
                                ],
                                "properties": {
                                    "immatriculation": {
                                        "type": "string",
                                        "example": "BENI-123"
                                    },
                                    "marque": {
                                        "type": "string",
                                        "example": "Toyota"
                                    },
                                    "modele": {
                                        "type": "string",
                                        "example": "Coaster"
                                    },
                                    "couleur": {
                                        "type": "string",
                                        "example": "Blanc"
                                    },
                                    "chauffeur_id": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bus enregistré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "bus": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "immatriculation": {
                                                    "type": "string"
                                                },
                                                "marque": {
                                                    "type": "string"
                                                },
                                                "modele": {
                                                    "type": "string"
                                                },
                                                "couleur": {
                                                    "type": "string"
                                                },
                                                "chauffeur_id": {
                                                    "type": "integer"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}